python/pyproject.toml
[project]
name = "voxels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "compliance-checker>=5.1.1",
    "geopandas>=1.0.1",
    "rioxarray>=0.18.2",
    "xarray>=2025.1.2",
]
